feat: make listings page home page
